+++ test-expected/sequelize/dumper-output/export.expected.route.js
@@ -1,0 +1,60 @@
+const express = require('express');
+const { PermissionMiddlewareCreator } = require('forest-express-sequelize');
+const { modelExport } = require('../models');
+
+const router = express.Router();
+const permissionMiddlewareCreator = new PermissionMiddlewareCreator('modelExport');
+
+// This file contains the logic of every route in Forest Admin for the collection modelExport:
+// - Native routes are already generated but can be extended/overriden - Learn how to extend a route here: https://docs.forestadmin.com/documentation/v/v6/reference-guide/routes/extend-a-route
+// - Smart action routes will need to be added as you create new Smart Actions - Learn how to create a Smart Action here: https://docs.forestadmin.com/documentation/v/v6/reference-guide/actions/create-and-manage-smart-actions
+
+// Create a Export
+router.post('/modelExport', permissionMiddlewareCreator.create(), (request, response, next) => {
+  // Learn what this route does here: https://docs.forestadmin.com/documentation/v/v6/reference-guide/routes/default-routes#create-a-record
+  next();
+});
+
+// Update a Export
+router.put('/modelExport/:recordId', permissionMiddlewareCreator.update(), (request, response, next) => {
+  // Learn what this route does here: https://docs.forestadmin.com/documentation/v/v6/reference-guide/routes/default-routes#update-a-record
+  next();
+});
+
+// Delete a Export
+router.delete('/modelExport/:recordId', permissionMiddlewareCreator.delete(), (request, response, next) => {
+  // Learn what this route does here: https://docs.forestadmin.com/documentation/v/v6/reference-guide/routes/default-routes#delete-a-record
+  next();
+});
+
+// Get a list of Exports
+router.get('/modelExport', permissionMiddlewareCreator.list(), (request, response, next) => {
+  // Learn what this route does here: https://docs.forestadmin.com/documentation/v/v6/reference-guide/routes/default-routes#get-a-list-of-records
+  next();
+});
+
+// Get a number of Exports
+router.get('/modelExport/count', permissionMiddlewareCreator.list(), (request, response, next) => {
+  // Learn what this route does here: https://docs.forestadmin.com/documentation/v/v6/reference-guide/routes/default-routes#get-a-number-of-records
+  next();
+});
+
+// Get a Export
+router.get('/modelExport/:recordId', permissionMiddlewareCreator.details(), (request, response, next) => {
+  // Learn what this route does here: https://docs.forestadmin.com/documentation/v/v6/reference-guide/routes/default-routes#get-a-record
+  next();
+});
+
+// Export a list of Exports
+router.get('/modelExport.csv', permissionMiddlewareCreator.export(), (request, response, next) => {
+  // Learn what this route does here: https://docs.forestadmin.com/documentation/v/v6/reference-guide/routes/default-routes#export-a-list-of-records
+  next();
+});
+
+// Delete a list of Exports
+router.delete('/modelExport', permissionMiddlewareCreator.delete(), (request, response, next) => {
+  // Learn what this route does here: https://docs.forestadmin.com/documentation/v/v6/reference-guide/routes/default-routes#delete-a-list-of-records
+  next();
+});
+
+module.exports = router;